fixed short flag bug

+++ src/main.zig
@@ -13,7 +13,7 @@ pub fn main() void {
 
     for (args.items) |item| {
         if (item.value == .Option) {
-            std.debug.print("{s} : {} ({s})\n", .{ item.value.Option.name, item.value.Option.int(i32) catch -1, item.value.typeAsString() });
+            std.debug.print("{s} : {} ({s})\n", .{ item.value.Option.name, item.value.Option.float(f32) catch -1, item.value.typeAsString() });
         } else {
             std.debug.print("{s} ({s})\n", .{ item.asString(), item.value.typeAsString() });
         }
